⚡️ Feat: 회원가입 - 전문가 카테고리 설정 추가
Feat: 회원가입 - 전문가 카테고리 설정 추가
