+++ MYworkHER/accounts/templates/accounts/signup/step7.html
@@ -1,0 +1,21 @@
+{% load static %}
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <form method="POST">
+  {% csrf_token %}
+  <label>상세 분야를 선택해주세요</label>
+  <select name="subcategory" required>
+    {% for sub in subcategories %}
+      <option value="{{ sub }}">{{ sub }}</option>
+    {% endfor %}
+  </select>
+  <button type="submit">확인</button>
+</form>
+</body>
+</html>
