+++ MYworkHER/category/templates/category/category_expert.html
@@ -28,7 +28,7 @@
     {% endif %}
 
         <div style="display: none;" class="subcategory-item">
-          <a href="{% url 'experts:expert_category' %}?category={{ key }}&subcategory={{ sub.name }}">
+          <a href="{% url 'category:expert_category' %}?category={{ key }}&subcategory={{ sub.name }}">
             <span class="subName"> {{ sub.name }}</span> <span class="subCount">전문가 {{ sub.count }}명</span>
           </a>
         </div>
