+++ MYworkHER/chats/templates/chats/chat-list.html
@@ -67,7 +67,7 @@
     <img src="{% static 'img/empty-chat.svg' %}" alt="채팅 없음" class="chat-empty-img" />
     <p class="chat-empty-text">아직 진행 중인 1:1 상담 내역이 없어요!</p>
     <div class="chat-empty-buttons">
-      <a href="{% url 'experts:category_list' %}">카테고리 둘러보기 ></a>
+      <a href="{% url 'category:category_list' %}">카테고리 둘러보기 ></a>
       <a href="{% url 'experts:expert_list' %}">전문가 찾기 ></a>
     </div>
   </div>
