+++ MYworkHER/experts/templates/experts/expert_search.html
@@ -22,7 +22,7 @@
         <input class="Search" type="text" name="keyword" value="{{ keyword }}" placeholder="전문가명, 한줄소개, 키워드 검색">
       </div>
       <button type="button" class="SearchCategory">
-        <a href="{% url 'experts:category_list' %}">카테고리로 찾기 ></a>
+        <a href="{% url 'category:category_list' %}">카테고리로 찾기 ></a>
       </button>
     </div>
 
@@ -57,7 +57,7 @@
             <label class="custom-checkbox">
               <input class="checkbox" type="checkbox" name="region" value="{{ region }}" {% if region in selected_region %}checked{% endif %}>
             <span class="checkmark"></span>
-            <span class="checkbox-label">{{region}}</span>
+            <span class="checkbox-label">{{ region }}</span>
             </label>
       {% endfor %}
       </div>
