+++ MYworkHER/matching/templates/matching/main.html
@@ -18,7 +18,7 @@
                     <a href="{% url 'matching:main' %}" class="navbar-logo">
                         <img src="{% static 'img/logo.svg' %}" alt="로고" />
                     </a>
-                    <a href="{% url 'experts:category_list' %}" class="navbar-item">카테고리</a>
+                    <a href="{% url 'category:category_list' %}" class="navbar-item">카테고리</a>
                     <a href="{% url 'experts:expert_search' %}" class="navbar-item">전문가 찾기</a>
                     {% if request.user.is_authenticated %}
                         <a href="{% url 'chats:chat_list' %}" class="navbar-item">1&#58;1상담
@@ -54,7 +54,7 @@
         </div>
         <div class="Category">
             <div class="C1">
-                <a href="{% url 'experts:expert_category' %}?category=appliance">
+                <a href="{% url 'category:expert_category' %}?category=appliance">
                 <img
                     src="{% static 'img/main_C1.svg' %}"
                     alt="repair"
@@ -63,7 +63,7 @@
                 </a>
             </div>
             <div class="C2">
-                <a href="{% url 'experts:expert_category' %}?category=health">
+                <a href="{% url 'category:expert_category' %}?category=health">
                 <img
                     src="{% static 'img/main_C2.svg' %}"
                     alt="sports"
@@ -72,7 +72,7 @@
                 </a>
             </div>
             <div class="C3">
-                <a href="{% url 'experts:expert_category' %}?category=business"">
+                <a href="{% url 'category:expert_category' %}?category=business"">
                 <img
                     src="{% static 'img/main_C3.svg' %}"
                     alt="consulting"
@@ -81,7 +81,7 @@
                 </a>
             </div>
             <div class="C4">
-                <a href="{% url 'experts:expert_category' %}?category=lifestyle">
+                <a href="{% url 'category:expert_category' %}?category=lifestyle">
                 <img
                     src="{% static 'img/main_C4.svg' %}"
                     alt="life"
