+++ MYworkHER/templates/components/navbar.html
@@ -5,13 +5,8 @@
       <a href="{% url 'matching:main' %}" class="navbar-logo">
         <img src="{% static 'img/logo.svg' %}" alt="로고" />
       </a>
-      <!-- <a href="{% url 'experts:category_list' %}" class="navbar-item">카테고리</a>
-      <a href="{% url 'experts:expert_list' %}" class="navbar-item">전문가 찾기</a>
-      <a href="{% url 'chats:chat_list' %}" class="navbar-item">1&#58;1상담<span class="alarm-dot"></span></a>
-      <a href="{% url 'accounts:mypage' %}" class="navbar-item active">마이페이지</a> -->
-      
 
-<a href="{% url 'experts:category_list' %}" class="navbar-item {% if '/category' in request.path or request.path == '/experts/' %}active{% endif %}">카테고리</a>
+<a href="{% url 'category:category_list' %}" class="navbar-item {% if '/category' in request.path or request.path == '/category/' %}active{% endif %}">카테고리</a>
 <a href="{% url 'experts:expert_list' %}" class="navbar-item {% if '/experts/list' in request.path or '/experts/search' in request.path  %}active{% endif %}">전문가 찾기</a>
     
 <a href="{% url 'chats:chat_list' %}" class="navbar-item {% if '/chats' in request.path %}active{% endif %}">
